+++ src/components/Menu/MenuItem/MenuItem.jsx
@@ -1,22 +1,21 @@
-import React, { useCallback, useRef, useEffect, useState, useLayoutEffect, useReducer } from "react";
+import React, { useCallback, useRef, useEffect } from "react";
 
 import PropTypes from "prop-types";
 import isFunction from "lodash/isFunction";
 import cx from "classnames";
 
 import Icon from "../../Icon/Icon";
-
 import DropdownChevronRight from "../../Icon/Icons/components/DropdownChevronRight";
-import useKeyEvent from "../../../hooks/useKeyEvent";
+import DialogContentContainer from "../../DialogContentContainer/DialogContentContainer";
+
 import useMergeRefs from "../../../hooks/useMergeRefs";
 import useIsOverflowing from "../../../hooks/useIsOverflowing";
 import useIsMouseOver from "../../../hooks/useIsMouseOver";
-import useIsMouseEnter from "../../../hooks/useIsMouseEnter";
-import useForceUpdate from "../../../hooks/useForceUpdate";
-import usePrevious from "../../../hooks/usePrevious";
-
 import usePopover from "../../../hooks/usePopover";
-import DialogContentContainer from "../../DialogContentContainer/DialogContentContainer";
+
+import useMenuItemMouseEvents from "./hooks/useMenuItemMouseEvents";
+import useMenuItemKeyboardEvents from "./hooks/useMenuItemKeyboardEvents";
+
 import "./MenuItem.scss";
 
 const MenuItem = ({
@@ -40,6 +39,7 @@ const MenuItem = ({
   const isSubMenuOpen = !!children && isActive && hasOpenSubMenu;
   const hasChildren = !!children;
   const shouldShowSubMenu = hasChildren && isParentMenuVisible && isSubMenuOpen;
+  const submenuChild = children && React.Children.only(children);
 
   const ref = useRef(null);
   const titleRef = useRef();
@@ -49,72 +49,33 @@ const MenuItem = ({
   const popperElement = popperElementRef.current;
   const referenceElement = referenceElementRef.current;
   const childElement = childRef.current;
+
   const isTitleHoveredHovered = useIsMouseOver({ ref: titleRef });
   const isTitleHoveredAndOverflowing = useIsOverflowing({ ref: isTitleHoveredHovered && titleRef });
-  const isMouseEnter = useIsMouseEnter({ ref });
-  const forceUpdate = useForceUpdate();
 
   const { styles, attributes } = usePopover(referenceElement, popperElement, {
     isOpen: isSubMenuOpen
   });
 
-  const prevIsMouseEnter = usePrevious(isMouseEnter);
-
-  useEffect(() => {
-    if (!isMouseEnter) return;
-    if (isMouseEnter === prevIsMouseEnter) return;
-
-    if (!isActive) {
-      setActiveItemIndex(index);
-      if (hasChildren) {
-        setSubMenuIsOpenByIndex(index, true);
-      } else {
-        resetOpenSubMenuIndex();
-      }
-    }
-
-    if (isActive) {
-      setSubMenuIsOpenByIndex(index, !!isMouseEnter);
-    }
-  }, [
+  useMenuItemMouseEvents(
+    ref,
     resetOpenSubMenuIndex,
-    prevIsMouseEnter,
-    isMouseEnter,
     setSubMenuIsOpenByIndex,
     isActive,
     setActiveItemIndex,
     index,
-    forceUpdate,
     hasChildren
-  ]);
+  );
 
-  const onClickCallback = useCallback(
-    event => {
-      if (!isActive && !isMouseEnter) return;
-
-      if (isActive && hasChildren) {
-        setActiveItemIndex(index);
-        setSubMenuIsOpenByIndex(index, true);
-        return;
-      }
-
-      if (shouldShowSubMenu) return;
-
-      if (onClick && !disabled && isActive) {
-        onClick(event);
-      }
-    },
-    [
-      onClick,
-      isMouseEnter,
-      disabled,
-      isActive,
-      index,
-      setActiveItemIndex,
-      hasChildren,
-      shouldShowSubMenu,
-      setSubMenuIsOpenByIndex
-    ]
+  const { onClickCallback } = useMenuItemKeyboardEvents(
+    onClick,
+    disabled,
+    isActive,
+    index,
+    setActiveItemIndex,
+    hasChildren,
+    shouldShowSubMenu,
+    setSubMenuIsOpenByIndex
   );
 
   useEffect(() => {
@@ -125,11 +86,6 @@ const MenuItem = ({
     }
   }, [shouldShowSubMenu, childElement]);
 
-  useKeyEvent({
-    keys: ["Enter", "ArrowRight"],
-    callback: onClickCallback
-  });
-
   const closeSubMenu = useCallback(() => {
     setSubMenuIsOpenByIndex(index, false);
     focusParentMenu && focusParentMenu();
@@ -137,10 +93,8 @@ const MenuItem = ({
 
   const mergedRef = useMergeRefs({ refs: [ref, referenceElementRef] });
 
-  const submenuChild = children && React.Children.only(children);
-
   const renderSubMenuIconIfNeeded = () => {
-    if (!children) return null;
+    if (!hasChildren) return null;
 
     return (
       <div className="monday-style-menu-item__sub_menu_icon-wrapper">
@@ -200,6 +154,7 @@ const MenuItem = ({
 
       <div
         style={{ ...styles.popper, visibility: shouldShowSubMenu ? "visible" : "hidden" }}
+        // eslint-disable-next-line react/jsx-props-no-spreading
         {...attributes.popper}
         className="monday-style-menu-item__popover"
         ref={popperElementRef}
